split makeArray into separate files/test files
