+++ tdd/scripts/bubblesort.js
@@ -1,17 +1,4 @@
-
-
-let makeArray = () => {
-  let unsortedArray = [];
-  let numElements = 10;
-  let max  = 100;
-  let min = -100;
-    for (let i = 0; i < numElements; i++) {
-      unsortedArray.push(Math.floor(Math.random() * (max - (min))) + (min))
-    }
-
-  console.log(unsortedArray);
-  return unsortedArray;
-}
+import makeArray from '../scripts/makeArray'
 
 let bubblesort = sortedArray => {
   for (let i = 0; i < sortedArray.length - 1; i++) {
@@ -28,4 +15,4 @@ let bubblesort = sortedArray => {
 let unsortedArray = makeArray();
 bubblesort(unsortedArray);
 
-module.exports = bubblesort, makeArray;
+module.exports = bubblesort;
