+++ tdd/tests/bubblesort-test.js
@@ -1,25 +1,6 @@
 import { assert } from 'chai';
 import { expect } from 'chai';
 import bubblesort from '../scripts/bubblesort'
-import makeArray from '../scripts/bubblesort'
-
-describe('the function makeArray', () => {
-  it('should be a function', () => {
-    assert.isFunction(makeArray)
-  })
-
-  it.skip('should make an array', () => {
-    bubblesort([1, 2, 3])
-    expect(Array.isArray(makeArray()));
-  })
-
-  it.skip('should make an array with the desired length', () => {
-    let unsortedArray = makeArray();
-
-    // console.log(unsortedArray);
-    expect(unsortedArray).to.have.lengthOf(10);
-  })
-})
 
 describe('the function bubblesort', () => {
   it('should be a function', () => {
